+++ src/app/(routes)/(app)/[user]/[repository]/page.tsx
@@ -18,7 +18,7 @@ import { CounterBadge } from "~/app/(components)/counter-badge";
 import { Button } from "~/app/(components)/button";
 import { Markdown } from "~/app/(components)/markdown/markdown";
 import Link from "next/link";
-import { Cache } from "~/app/(components)/cache";
+import { Cache } from "~/app/(components)/cache/cache";
 import { Skeleton } from "~/app/(components)/skeleton";
 
 // utils
@@ -440,6 +440,7 @@ async function ReadmeContent(props: {
           id={CacheKeys.readme(props.user, props.repository)}
           updatedAt={updatedAt}
           ttl={THIRTY_MINUTES_IN_SECONDS}
+          bypassInDEV={false}
         >
           <Markdown
             linkHeaders
